+++ packages/expo-router/src/global-state/routing.ts
@@ -106,18 +106,18 @@ function rewriteNavigationStateToParams(
   state?: { routes: ResultState['routes'] },
   params: NavigationParams = {}
 ) {
-  if (!state) return params;
+  if (!state) return { ...params };
   // We Should always have at least one route in the state
   const lastRoute = state.routes.at(-1)!;
   params.screen = lastRoute.name;
   // Weirdly, this always needs to be an object. If it's undefined, it won't work.
-  params.params = lastRoute.params ?? {};
+  params.params = { ...lastRoute.params } ?? {};
 
   if (lastRoute.state) {
     rewriteNavigationStateToParams(lastRoute.state, params.params);
   }
 
-  return params;
+  return { ...params };
 }
 
 function getNavigatePushAction(state: ResultState) {